hasta pedidos navidad v1

+++ PIGGISWS/Models/DTOs/PedidoNavDetalle.cs
@@ -7,7 +7,7 @@ public class PedidoNavDetalle
         public int ID_PEDIDO_DET { get; set; }
         public int ID_EMPRESA { get; set; }
         public int ID_PEDIDO_FK { get; set; }
-        public int PRO_CODIGO { get; set; }
+        public decimal PRO_CODIGO { get; set; }
         public string PRO_NOMBRE { get; set; } = string.Empty;
         public decimal CANTIDAD { get; set; }
         public decimal PRECIO { get; set; }
